requirements.txt
openai==0.28.1
python-dotenv~=1.0.0
wikipedia
tiktoken
neo4j
streamlit~=1.29.0
sentence_transformers==2.2.2
Pillow
fastapi~=0.104.1
PyPDF2~=3.0.1
torch==2.0.1
pydantic~=2.5.2
uvicorn
sse-starlette
boto3
pdfplumber
langchain~=0.0.350